+++ flutter/ios/Classes/SentryFlutterPluginApple.swift
@@ -239,7 +239,7 @@ public class SentryFlutterPluginApple: NSObject, FlutterPlugin {
 
        if didReceiveDidBecomeActiveNotification &&
             (PrivateSentrySDKOnly.options.enableAutoSessionTracking ||
-             PrivateSentrySDKOnly.options.enableWatchdogTerminationsTracking) {
+             PrivateSentrySDKOnly.options.enableWatchdogTerminationTracking) {
             // We send a SentryHybridSdkDidBecomeActive to the Sentry Cocoa SDK, so the SDK will mimics
             // the didBecomeActiveNotification notification. This is needed for session and OOM tracking.
            NotificationCenter.default.post(name: Notification.Name("SentryHybridSdkDidBecomeActive"), object: nil)
@@ -319,8 +319,8 @@ public class SentryFlutterPluginApple: NSObject, FlutterPlugin {
             options.maxCacheItems = maxCacheItems
         }
 
-        if let enableWatchdogTerminationsTracking = arguments["enableWatchdogTerminationsTracking"] as? Bool {
-            options.enableWatchdogTerminationsTracking = enableWatchdogTerminationsTracking
+        if let enableWatchdogTerminationTracking = arguments["enableWatchdogTerminationTracking"] as? Bool {
+            options.enableWatchdogTerminationTracking = enableWatchdogTerminationTracking
         }
 
         if let sendClientReports = arguments["sendClientReports"] as? Bool {